Adds a new public API, `DittoTools.requestLogExport()`, for programmatically triggering the export of device logs to the Ditto Portal.
This new function encapsulates the logic for updating the `__small_peer_info` collection, which was previously implemented directly in the `ExportLogsToPortal` composable. The composable is refactored to use this new API.
